+++ monkey/monkey_island/cc/services/authentication_service/flask_resources/__init__.py
@@ -4,3 +4,4 @@ from .login import Login
 from .logout import Logout
 from .register_resources import register_resources
 from .agent_otp import AgentOTP
+from .agent_otp_login import AgentOTPLogin