+++ monkey/monkey_island/cc/services/authentication_service/flask_resources/register_resources.py
@@ -4,6 +4,7 @@ from common import DIContainer
 
 from ..authentication_facade import AuthenticationFacade
 from .agent_otp import AgentOTP
+from .agent_otp_login import AgentOTPLogin
 from .login import Login
 from .logout import Logout
 from .register import Register
@@ -20,3 +21,4 @@ def register_resources(api: flask_restful.Api, container: DIContainer):
     api.add_resource(Login, *Login.urls, resource_class_args=(authentication_facade,))
     api.add_resource(Logout, *Logout.urls, resource_class_args=(authentication_facade,))
     api.add_resource(AgentOTP, *AgentOTP.urls)
+    api.add_resource(AgentOTPLogin, *AgentOTPLogin.urls)